Updated and restructured reader
