+++ code/ModelConversionApp/Reader/ILoader.cs
@@ -1,0 +1,8 @@
+﻿using ModelConversionApp.Models.LakeDatabase;
+
+namespace ModelConversionApp.Reader;
+
+internal interface ILoader
+{
+    internal (List<Table> tables, List<Relationship> relationships) LoadModel();
+}
