--- code/ModelConversionApp/Reader/Loader/ILoader.cs
@@ -1,6 +1,0 @@
-﻿namespace ModelConversionApp.Reader.Loader;
-
-internal interface ILoader
-{
-    internal void LoadModel();
-}
